Only b64encode ssl_ca if it needs it
The description of the ssl_ca config option in config.yaml states
that the certificate should be base64 encoded. But if it is then
the charm b64encodes it again when sending the ca down the client
relations. This change gates encoding the ca on whether it is
already encoded.

Change-Id: I9828d7567fd7f04cd0d80229ea1ff1275ea4269e
Closes-Bug: #1798066