Pre-focal-enablement for rabbitmq-server charm
This patch updates the rabbitmq charm to work on focal as part of other
test bundles, but does not update the charm for testing on focal. Hence
focal/ussuri bundles are not added and the metadata.yaml does NOT
include focal in this patchset.

A further patchset will be submitted that adds the focal testing to the
charm.  In particular, this patchset is required for the cinder focal
enablement.

Change-Id: Ia239b7c2f0ed2383e220cf0fa4ade443149a3b32