Forget cluster node as an action
There are race conditions in which the forget_cluster_node can get
executed against a node that is in the process of joining the cluster.

This change moves forget_cluster_node to an action that can be
performed by the administrator. The asses_cluster_status has been
updated to check for departed nodes and set status pointing toward the
use of the new forget-cluster-node action.

Closes-Bug: #1818260
Change-Id: I64bcdb9811a3816b394395fac19f5af5cc9f9006